requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.7
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
requests>=2.31.0
